requirements.txt
aiocache==0.12.3
aiodns==3.6.1
aiofiles==25.1.0
APScheduler==3.10.4
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-client-cache==0.13.0
aiolimiter==1.2.1
aiosignal==1.4.0
aiosqlite==0.17.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.10.0
asyncio-throttle==1.0.2
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blis==1.3.3
brotli==1.2.0
curl-cffi==0.13.0
faker==34.0.2
fastapi==0.121.1
ffmpeg-python==0.2.0
filelock==3.20.0
flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
fsspec==2025.10.0
google-generativeai==0.8.6
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
imagehash==4.3.2
itsdangerous==2.2.0
jinja2==3.1.6
lxml==6.0.2
mako==1.3.10
markdown==3.10
markupsafe==3.0.3
multidict==6.7.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
passlib==1.7.4
pillow==11.3.0
playwright==1.55.0
pluggy==1.6.0
prompt-toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.1.3
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic-core==2.41.5
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
scikit-learn==1.8.0
scipy==1.16.3
selenium==4.39.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.44
sse-starlette==3.0.4
starlette==0.49.3
static-ffmpeg==2.13
telethon==1.42.0
tqdm==4.67.1
typing-extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
yt-dlp==2025.12.8
zipstream-ng==1.9.0
